+++ src/collections.ts
@@ -17,17 +17,25 @@ import './rules';
 import {registry} from './registry';
 import {RuleCollection} from './rule';
 
-registry.register(new RuleCollection(
-    'polymer-2',
-    `Rules for projects that use Polymer 2.x`,
-    ['dom-module-invalid-attrs', 'style-into-template', 'undefined-elements']));
+registry.register(
+    new RuleCollection('polymer-2', `Rules for projects that use Polymer 2.x`, [
+      'dom-module-invalid-attrs',
+      'style-into-template',
+      'undefined-elements',
+      'unbalanced-polymer-delimiters',
+    ]));
 
 registry.register(new RuleCollection(
     'polymer-2-hybrid',
     `Rules for projects that are compatible with either Polymer 1.x or 2.x
 
 Will warn about use of deprecated Polymer 1.x features or brand new features in Polymer 2.x`,
-    ['dom-module-invalid-attrs', 'style-into-template', 'undefined-elements']));
+    [
+      'dom-module-invalid-attrs',
+      'style-into-template',
+      'undefined-elements',
+      'unbalanced-polymer-delimiters',
+    ]));
 
 registry.register(new RuleCollection(
     'polymer-1',
@@ -35,4 +43,7 @@ registry.register(new RuleCollection(
 
 For projects that are ready to start transitioning to Polymer 2.0 see polymer-2-hybrid.
 `,
-    ['undefined-elements']));
+    [
+      'undefined-elements',
+      'unbalanced-polymer-delimiters',
+    ]));